allbachelor/requirements.txt
Django
numpy
pandas
scipy
bootstrap3
certifi
django-markdown
Pillow
matplotlib
pymysql
mysqlclient
scikit_learn
sklearn
